+++ loadgen/cmd/otelbench/flags.go
@@ -43,12 +43,16 @@ var Config struct {
 	Logs    bool
 	Metrics bool
 	Traces  bool
+	Mixed   bool
 
 	ExporterOTLP     bool
 	ExporterOTLPHTTP bool
 
 	ConcurrencyList []int
 	Shuffle         bool
+	TracesDataPath  string
+	MetricsDataPath string
+	LogsDataPath    string
 
 	Telemetry TelemetryConfig
 }
@@ -132,6 +136,11 @@ func Init() error {
 	flag.BoolVar(&Config.Logs, "logs", true, "benchmark logs")
 	flag.BoolVar(&Config.Metrics, "metrics", true, "benchmark metrics")
 	flag.BoolVar(&Config.Traces, "traces", true, "benchmark traces")
+	flag.BoolVar(&Config.Mixed, "mixed", true, "benchmark mixed signals, i.e. logs, metrics and traces at the same time")
+
+	flag.StringVar(&Config.TracesDataPath, "traces-data-path", "", "path to traces data file (e.g. traces.json). If empty, embedded data will be used.")
+	flag.StringVar(&Config.MetricsDataPath, "metrics-data-path", "", "path to metrics data file (e.g. metrics.json). If empty, embedded data will be used.")
+	flag.StringVar(&Config.LogsDataPath, "logs-data-path", "", "path to logs data file (e.g. logs.json). If empty, embedded data will be used.")
 
 	flag.BoolVar(&Config.Shuffle, "shuffle", false, "shuffle the order of benchmarks. This is useful for concurrent runs.")
 
@@ -333,4 +342,21 @@ func SetConcurrency(concurrency int) (configFiles []string) {
 	return setsToConfigs([]string{
 		fmt.Sprintf("receivers.loadgen.concurrency=%d", concurrency),
 	})
+}
+
+// SetDataPaths returns a config override to set the data paths for loadgenreceiver.
+// Configuration options for `traces_data_path`, `metrics_data_path`, and `logs_data_path` will update
+// the existing 'jsonl_file' option for each signal.
+func SetDataPaths(tracesPath, metricsPath, logsPath string) []string {
+	var sets []string
+	if tracesPath != "" {
+		sets = append(sets, fmt.Sprintf("receivers.loadgen.traces.jsonl_file=%q", tracesPath))
+	}
+	if metricsPath != "" {
+		sets = append(sets, fmt.Sprintf("receivers.loadgen.metrics.jsonl_file=%q", metricsPath))
+	}
+	if logsPath != "" {
+		sets = append(sets, fmt.Sprintf("receivers.loadgen.logs.jsonl_file=%q", logsPath))
+	}
+	return setsToConfigs(sets)
 }